+++ packages/column-views/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.2] - 2025-12-04
+
+- Fix a bug with unit deselection
+- Fix missed updates in state management code
+- Add a 'minimal' option to `unconformityLabels`
+- Reduce precision of gap age labels
+- Improvements to stories
+
 ## [2.2.1] - 2025-11-29
 
 - Start unifying state management components